requirements.txt
# Core ML libraries
tensorflow>=2.12.0
keras>=2.12.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Image processing
Pillow>=8.3.0
opencv-python>=4.5.0

# Data visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Web application
streamlit>=1.25.0
gradio>=3.35.0

# Data handling
kaggle>=1.5.12
requests>=2.25.0

# Progress bars and utilities
tqdm>=4.62.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0